Implemented NodeFromWorldspace
When providing a Vector3 to NodeFromWorldpsace it will add a node connected to the closest node (or two nodes if  it's in between two nodes). It does linecasting to check if it's able to connect to it. Also added in a part that redoes the first cabin's nodes after they all other nodes are made as there was some jank happening there. Check the TODO in NodeManager under the Build Links function.