Path Request Manager
Fixed Up some issues with the NodeManager script not properly adding to the node list and not adding the neighbours to the new nodes themselves.

Questioning whether or not the new nodes need to be inserted into the grid or not? TODO clean up path request manager as it wont work with 2 AIs requesting a path at themoment