NPCs now walk to nodes
For some reason Unity's Angle system is based off of 0,0 rather than the two objects themselves. That was causing the issue of the NPC walking towards the wrong direction because they were always facing 0,0 rather than the proper direction. I've replaced the angle with a InverseTransformPoint which will output the provided vector to Local vector relative to the NPC so if it's negative they are to the left if it is positive they are to the right. Then sign it to be either -1 or 1 and use that.