Elevator System start
Beginnings of the elevator system (to replace the use of doors as elevators (maybe maybe not). As said in the comment this is probably going to force me to rewrite a lot of the Cabin and Pathfinding as if there are more than just 2 levels the pathfinder is going to need to know which level it is on. This could be done by just adding an int to the Node object and assigning that when building. when creating a node that isn't a node or a door just assign it to it's closest neighbor (?). Check Comments in Elevator.cs for more info